split up gamevars a into smaller headers
