+++ src/resources/mgs2_equipment_enums.hpp
@@ -1,0 +1,161 @@
+#pragma once
+
+enum MGS2WeaponIndex : uint8_t
+{
+    MGS2_WEAPON_INDEX_NONE = 0x0,
+    MGS2_WEAPON_INDEX_M9 = 0x1,
+    MGS2_WEAPON_INDEX_USP = 0x2,
+    MGS2_WEAPON_INDEX_SOCOM = 0x3,
+    MGS2_WEAPON_INDEX_PSG1 = 0x4,
+    MGS2_WEAPON_INDEX_RGB6 = 0x5,
+    MGS2_WEAPON_INDEX_NIKITA = 0x6,
+    MGS2_WEAPON_INDEX_STINGER = 0x7,
+    MGS2_WEAPON_INDEX_CLAYMORE = 0x8,
+    MGS2_WEAPON_INDEX_C4 = 0x9,
+    MGS2_WEAPON_INDEX_CHAFF_GRENADE = 0xA,
+    MGS2_WEAPON_INDEX_STUN_GRENADE = 0xB,
+    MGS2_WEAPON_INDEX_D_MIC = 0xC,
+    MGS2_WEAPON_INDEX_HIGH_FREQUENCY_BLADE = 0xD,
+    MGS2_WEAPON_INDEX_COOLANT = 0xE,
+    MGS2_WEAPON_INDEX_AKS74U = 0xF,
+    MGS2_WEAPON_INDEX_MAGAZINE = 0x10,
+    MGS2_WEAPON_INDEX_GRENADE = 0x11,
+    MGS2_WEAPON_INDEX_M4 = 0x12,
+    MGS2_WEAPON_INDEX_PSG1T = 0x13,
+    MGS2_WEAPON_INDEX_D_MIC_ZOOMED = 0x14,
+    MGS2_WEAPON_INDEX_BOOK = 0x15,
+
+    MGS2_WEAPON_INDEX_MAX_WEAPONS = 0x16,
+
+    MGS2_WEAPON_INDEX_PLAYER_ATTACK = 0x17,
+    MGS2_WEAPON_INDEX_BLADE_FAINT = 0x18,
+    MGS2_WEAPON_INDEX_BLADE_STAB = 0x19,
+    MGS2_WEAPON_INDEX_BLADE_GUARD = 0x1A,
+    MGS2_WEAPON_INDEX_BOX_REMOVE = 0x1B,
+    MGS2_WEAPON_INDEX_KICK1 = 0x1C,
+};
+
+enum MGS2WeaponFlags : uint64_t
+{
+    MGS2_WEAPON_NONE = (1ull << MGS2_WEAPON_INDEX_NONE),
+    MGS2_WEAPON_M9 = (1ull << MGS2_WEAPON_INDEX_M9),
+    MGS2_WEAPON_USP = (1ull << MGS2_WEAPON_INDEX_USP),
+    MGS2_WEAPON_SOCOM = (1ull << MGS2_WEAPON_INDEX_SOCOM),
+    MGS2_WEAPON_PSG1 = (1ull << MGS2_WEAPON_INDEX_PSG1),
+    MGS2_WEAPON_RGB6 = (1ull << MGS2_WEAPON_INDEX_RGB6),
+    MGS2_WEAPON_NIKITA = (1ull << MGS2_WEAPON_INDEX_NIKITA),
+    MGS2_WEAPON_STINGER = (1ull << MGS2_WEAPON_INDEX_STINGER),
+    MGS2_WEAPON_CLAYMORE = (1ull << MGS2_WEAPON_INDEX_CLAYMORE),
+    MGS2_WEAPON_C4 = (1ull << MGS2_WEAPON_INDEX_C4),
+    MGS2_WEAPON_CHAFF_GRENADE = (1ull << MGS2_WEAPON_INDEX_CHAFF_GRENADE),
+    MGS2_WEAPON_STUN_GRENADE = (1ull << MGS2_WEAPON_INDEX_STUN_GRENADE),
+    MGS2_WEAPON_D_MIC = (1ull << MGS2_WEAPON_INDEX_D_MIC),
+    MGS2_WEAPON_HIGH_FREQUENCY_BLADE = (1ull << MGS2_WEAPON_INDEX_HIGH_FREQUENCY_BLADE),
+    MGS2_WEAPON_COOLANT = (1ull << MGS2_WEAPON_INDEX_COOLANT),
+    MGS2_WEAPON_AKS74U = (1ull << MGS2_WEAPON_INDEX_AKS74U),
+    MGS2_WEAPON_MAGAZINE = (1ull << MGS2_WEAPON_INDEX_MAGAZINE),
+    MGS2_WEAPON_GRENADE = (1ull << MGS2_WEAPON_INDEX_GRENADE),
+    MGS2_WEAPON_M4 = (1ull << MGS2_WEAPON_INDEX_M4),
+    MGS2_WEAPON_PSG1T = (1ull << MGS2_WEAPON_INDEX_PSG1T),
+    MGS2_WEAPON_D_MIC_ZOOMED = (1ull << MGS2_WEAPON_INDEX_D_MIC_ZOOMED),
+    MGS2_WEAPON_BOOK = (1ull << MGS2_WEAPON_INDEX_BOOK),
+};
+
+enum MGS2ItemIndex : uint8_t
+{
+    MGS2_ITEM_INDEX_NONE = 0x0,  /// Unarmed / no item
+    MGS2_ITEM_INDEX_RATION = 0x1,
+    MGS2_ITEM_INDEX_DUMMY_SCOPE = 0x2,  /// Dummy binoculars
+    MGS2_ITEM_INDEX_MEDICINE = 0x3,
+    MGS2_ITEM_INDEX_STYPTIC = 0x4,  ///bandage
+    MGS2_ITEM_INDEX_DIAZEPAM = 0x5,  /// pentazemin
+    MGS2_ITEM_INDEX_UNIFORM = 0x6,  /// BDU
+    MGS2_ITEM_INDEX_BODY_ARMOR = 0x7,
+    MGS2_ITEM_INDEX_STEALTH = 0x8,
+    MGS2_ITEM_INDEX_MINE_DETECTOR = 0x9,
+
+    MGS2_ITEM_INDEX_BOMB_SENSOR_A = 0xA,
+    MGS2_ITEM_INDEX_BOMB_SENSOR_B = 0xB,
+    MGS2_ITEM_INDEX_NIGHT_VISION_GOGGLES = 0xC,
+    MGS2_ITEM_INDEX_THERMAL_GOGGLES = 0xD,
+    MGS2_ITEM_INDEX_SCOPE = 0xE,
+    MGS2_ITEM_INDEX_CAMERA = 0xF,
+    MGS2_ITEM_INDEX_CARDBOARD_BOX_A = 0x10,
+    MGS2_ITEM_INDEX_CIGARETTES = 0x11,
+    MGS2_ITEM_INDEX_ID_CARD = 0x12,
+    MGS2_ITEM_INDEX_SHAVER = 0x13,
+
+    MGS2_ITEM_INDEX_PHS = 0x14, /// Cell phone
+    MGS2_ITEM_INDEX_TANKER_CAMERA = 0x15,
+    MGS2_ITEM_INDEX_CARDBOARD_BOX_B = 0x16,
+    MGS2_ITEM_INDEX_CARDBOARD_BOX_C = 0x17,
+    MGS2_ITEM_INDEX_WET_CARDBOARD_BOX = 0x18,
+    MGS2_ITEM_INDEX_VIBRATION_SENSOR = 0x19,
+    MGS2_ITEM_INDEX_CARDBOARD_BOX_D = 0x1A,
+    MGS2_ITEM_INDEX_CARDBOARD_BOX_E = 0x1B,
+    MGS2_ITEM_INDEX_RAZOR = 0x1C, //cut body razor?
+    MGS2_ITEM_INDEX_SOCOM_SUPPRESSOR = 0x1D,
+
+    MGS2_ITEM_INDEX_AKS74U_SUPPRESSOR = 0x1E,
+    MGS2_ITEM_INDEX_DUMMY_TANKER_CAMERA = 0x1F,
+    MGS2_ITEM_INDEX_INFINITY_BANDANA = 0x20,
+    MGS2_ITEM_INDEX_DOG_TAG = 0x21,
+    MGS2_ITEM_INDEX_MO_DISC = 0x22,
+    MGS2_ITEM_INDEX_USP_SUPPRESSOR = 0x23,
+    MGS2_ITEM_INDEX_INFINITY_WIG = 0x24,
+    MGS2_ITEM_INDEX_WIG_A = 0x25, /// Wig A (infinite O2)
+    MGS2_ITEM_INDEX_WIG_B = 0x26, /// Wig B (infinite grip)
+    MGS2_ITEM_INDEX_WIG_C = 0x27, //?
+
+    MGS2_ITEM_INDEX_WIG_D = 0x28, //?
+
+    MGS2_ITEM_INDEX_MAX_ITEMS = 0x29,
+};
+
+enum MGS2ItemFlags : uint64_t
+{
+    MGS2_ITEM_FLAG_NONE = (1ull << MGS2_ITEM_INDEX_NONE),
+    MGS2_ITEM_FLAG_RATION = (1ull << MGS2_ITEM_INDEX_RATION),
+    MGS2_ITEM_FLAG_DUMMY_SCOPE = (1ull << MGS2_ITEM_INDEX_DUMMY_SCOPE),
+    MGS2_ITEM_FLAG_MEDICINE = (1ull << MGS2_ITEM_INDEX_MEDICINE),
+    MGS2_ITEM_FLAG_STYPTIC = (1ull << MGS2_ITEM_INDEX_STYPTIC),
+    MGS2_ITEM_FLAG_DIAZEPAM = (1ull << MGS2_ITEM_INDEX_DIAZEPAM),
+    MGS2_ITEM_FLAG_UNIFORM = (1ull << MGS2_ITEM_INDEX_UNIFORM),
+    MGS2_ITEM_FLAG_BODY_ARMOR = (1ull << MGS2_ITEM_INDEX_BODY_ARMOR),
+    MGS2_ITEM_FLAG_STEALTH = (1ull << MGS2_ITEM_INDEX_STEALTH),
+    MGS2_ITEM_FLAG_MINE_DETECTOR = (1ull << MGS2_ITEM_INDEX_MINE_DETECTOR),
+
+    MGS2_ITEM_FLAG_BOMB_SENSOR_A = (1ull << MGS2_ITEM_INDEX_BOMB_SENSOR_A),
+    MGS2_ITEM_FLAG_BOMB_SENSOR_B = (1ull << MGS2_ITEM_INDEX_BOMB_SENSOR_B),
+    MGS2_ITEM_FLAG_NIGHT_VISION_GOGGLES = (1ull << MGS2_ITEM_INDEX_NIGHT_VISION_GOGGLES),
+    MGS2_ITEM_FLAG_THERMAL_GOGGLES = (1ull << MGS2_ITEM_INDEX_THERMAL_GOGGLES),
+    MGS2_ITEM_FLAG_SCOPE = (1ull << MGS2_ITEM_INDEX_SCOPE),
+    MGS2_ITEM_FLAG_CAMERA = (1ull << MGS2_ITEM_INDEX_CAMERA),
+    MGS2_ITEM_FLAG_CARDBOARD_BOX_A = (1ull << MGS2_ITEM_INDEX_CARDBOARD_BOX_A),
+    MGS2_ITEM_FLAG_CIGARETTES = (1ull << MGS2_ITEM_INDEX_CIGARETTES),
+    MGS2_ITEM_FLAG_ID_CARD = (1ull << MGS2_ITEM_INDEX_ID_CARD),
+    MGS2_ITEM_FLAG_SHAVER = (1ull << MGS2_ITEM_INDEX_SHAVER),
+
+    MGS2_ITEM_FLAG_PHS = (1ull << MGS2_ITEM_INDEX_PHS),
+    MGS2_ITEM_FLAG_TANKER_CAMERA = (1ull << MGS2_ITEM_INDEX_TANKER_CAMERA),
+    MGS2_ITEM_FLAG_CARDBOARD_BOX_B = (1ull << MGS2_ITEM_INDEX_CARDBOARD_BOX_B),
+    MGS2_ITEM_FLAG_CARDBOARD_BOX_C = (1ull << MGS2_ITEM_INDEX_CARDBOARD_BOX_C),
+    MGS2_ITEM_FLAG_WET_CARDBOARD_BOX = (1ull << MGS2_ITEM_INDEX_WET_CARDBOARD_BOX),
+    MGS2_ITEM_FLAG_VIBRATION_SENSOR = (1ull << MGS2_ITEM_INDEX_VIBRATION_SENSOR),
+    MGS2_ITEM_FLAG_CARDBOARD_BOX_D = (1ull << MGS2_ITEM_INDEX_CARDBOARD_BOX_D),
+    MGS2_ITEM_FLAG_CARDBOARD_BOX_E = (1ull << MGS2_ITEM_INDEX_CARDBOARD_BOX_E),
+    MGS2_ITEM_FLAG_RAZOR = (1ull << MGS2_ITEM_INDEX_RAZOR),
+    MGS2_ITEM_FLAG_SOCOM_SUPPRESSOR = (1ull << MGS2_ITEM_INDEX_SOCOM_SUPPRESSOR),
+
+    MGS2_ITEM_FLAG_AKS74U_SUPPRESSOR = (1ull << MGS2_ITEM_INDEX_AKS74U_SUPPRESSOR),
+    MGS2_ITEM_FLAG_DUMMY_TANKER_CAMERA = (1ull << MGS2_ITEM_INDEX_DUMMY_TANKER_CAMERA),
+    MGS2_ITEM_FLAG_INFINITY_BANDANA = (1ull << MGS2_ITEM_INDEX_INFINITY_BANDANA),
+    MGS2_ITEM_FLAG_DOG_TAG = (1ull << MGS2_ITEM_INDEX_DOG_TAG),
+    MGS2_ITEM_FLAG_MO_DISC = (1ull << MGS2_ITEM_INDEX_MO_DISC),
+    MGS2_ITEM_FLAG_USP_SUPPRESSOR = (1ull << MGS2_ITEM_INDEX_USP_SUPPRESSOR),
+    MGS2_ITEM_FLAG_INFINITY_WIG = (1ull << MGS2_ITEM_INDEX_INFINITY_WIG),
+    MGS2_ITEM_FLAG_WIG_A = (1ull << MGS2_ITEM_INDEX_WIG_A),
+    MGS2_ITEM_FLAG_WIG_B = (1ull << MGS2_ITEM_INDEX_WIG_B),
+    MGS2_ITEM_FLAG_WIG_C = (1ull << MGS2_ITEM_INDEX_WIG_C),
+    MGS2_ITEM_FLAG_WIG_D = (1ull << MGS2_ITEM_INDEX_WIG_D),
+};
